requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-multipart==0.0.6
httpx==0.25.1